Rename: db->database, api->controller
