+++ run.py
@@ -1,4 +1,4 @@
-from app.app import app
+from app import app
 
 
 if __name__ == "__main__":
